+++ packages/hippo_dev/hippo_analysis/test/hippo_analysis_test.dart
@@ -1,0 +1,31 @@
+import 'dart:io';
+import 'dart:isolate';
+
+import 'package:hippo_analysis/hippo_analysis.dart';
+import 'package:test/test.dart';
+
+void main() {
+  test('YAML and Dart formatter policies stay aligned', () async {
+    final optionsUri = await Isolate.resolvePackageUri(
+      Uri.parse('package:hippo_analysis/base.yaml'),
+    );
+    final options = File.fromUri(optionsUri!).readAsStringSync();
+
+    expect(options, contains('page_width: $hippoFormatterPageWidth'));
+    expect(options, contains('trailing_commas: ${hippoFormatterTrailingCommas.name}'));
+  });
+
+  test('shared formatter uses automatic trailing commas and a 100-column page', () {
+    final formatted = createHippoDartFormatter().format('''
+void main() {
+  call('12345678901234567890123456789012345678901234567890', '12345678901234567890123456789012345678901234567890');
+  call('short',);
+}
+''');
+
+    expect(formatted, contains("  call(\n"));
+    expect(formatted, contains("    '12345678901234567890123456789012345678901234567890',\n"));
+    expect(formatted, contains("  );\n"));
+    expect(formatted, contains("  call('short');\n"));
+  });
+}